api tests + fixes + frontend skeleton

+++ backend/app/api/routes/users/controllers/delete_user.py
@@ -1,7 +1,11 @@
-from app.api.exceptions.exceptions import UnauthorizedException
+from sqlalchemy import select
+from fastapi import Response
+
+from app.config import Config
+from app.api.exceptions.exceptions import UnauthorizedException, CannotDeleteLastOwnerException, NotFoundException
 from app.api.schemas import SuccessResponse, User
-from app.constants import Permission
-from app.services.database.mysql.schemas.user import UsersTable
+from app.constants import Permission, UserRole
+from app.services.database.mysql.schemas.user import UsersTable, UserRow
 from app.services.database.mysql.service import MySQLService
 
 
@@ -11,12 +15,29 @@ class DeleteUserController:
         self.user_id = user_id
         self.me = me
 
-    def handle_request(self) -> SuccessResponse:
+    def handle_request(self, response: Response) -> SuccessResponse:
         if not self.me.role.has_permission(Permission.DELETE_USER):
             raise UnauthorizedException
 
         with MySQLService.get_session() as session:
+            if not session.get(UserRow, self.user_id):
+                raise NotFoundException
+
+            user_row = session.scalar(
+                select(
+                    UserRow.user_id
+                ).where(
+                    UserRow.role == UserRole.OWNER.value,
+                    UserRow.user_id != self.user_id
+                ).limit(1)
+            )
+            if not user_row:
+                raise CannotDeleteLastOwnerException
+
             UsersTable.delete_user(user_id=self.user_id, session=session)
             session.commit()
+
+        if self.user_id == self.me.user_id:
+            response.delete_cookie(key=Config.SESSION_COOKIE_KEY)
 
         return SuccessResponse()
